Updated App Styling
Updated app styling to have the selected tab icon show up as the selected gold (changed the --ion-color-primary-contrast variable in the variables.scss file to do so).
Updated the shared header to set the title color back to white after the change to the primary color contrast color.
Added a blank "Info" page and set up the routing and tab navigation for it.
Moved "tabs" folder into the "pages" folder.
Moved merged tabs routing with app routing and removed tabs module.